+++ WalletWasabi.Fluent/Controls/ReceiveAddressControl.axaml
@@ -25,7 +25,7 @@
     <Setter Property="Template">
       <ControlTemplate>
         <DockPanel>
-          <StackPanel DockPanel.Dock="Bottom" Spacing="30">
+          <StackPanel DockPanel.Dock="Bottom" Spacing="15">
 
             <StackPanel Orientation="Horizontal" HorizontalAlignment="Center" Spacing="8">
               <controls:PrivacyContentControl FontSize="{StaticResource FontSizeH5}">
