+++ WalletWasabi.Fluent/ViewModels/Wallets/Home/Tiles/BtcPriceTileViewModel.cs
@@ -1,22 +1,8 @@
-using ReactiveMarbles.PropertyChanged;
-using ReactiveUI;
 using WalletWasabi.Bridge;
 
 namespace WalletWasabi.Fluent.ViewModels.Wallets.Home.Tiles;
 
-public partial class BtcPriceTileViewModelDesign : ReactiveObject, IBtcPriceTileViewModel
-{
-	[AutoNotify] private decimal _usdPerBtcValue;
-
-	public BtcPriceTileViewModelDesign()
-	{
-		UsdPerBtc = this.WhenChanged(x => x.UsdPerBtcValue);
-	}
-
-	public IObservable<decimal> UsdPerBtc { get; }
-}
-
-public class BtcPriceTileViewModel : ActivatableViewModel
+public class BtcPriceTileViewModel : ActivatableViewModel, IBtcPriceTileViewModel
 {
 	public BtcPriceTileViewModel(IExchangeRateProvider exchangeRateProvider)
 	{
