+++ WalletWasabi.Fluent/ViewModels/Wallets/WalletViewModel.cs
@@ -36,6 +36,7 @@ public partial class WalletViewModel : WalletViewModelBase
 	[AutoNotify(SetterModifier = AccessModifier.Private)] private bool _isWalletBalanceZero;
 	[AutoNotify(SetterModifier = AccessModifier.Private)] private bool _isTransactionHistoryEmpty;
 	[AutoNotify(SetterModifier = AccessModifier.Private)] private bool _isSendButtonVisible;
+
 	private readonly IWallet _myWallet;
 	private readonly ExchangeRateProvider _exchangeRateProvider;
